move CustomExceptionHandlerOptions
