+++ src/WeihanLi.Web.Extensions/Middleware/CustomExceptionHandlerMiddleware.cs
@@ -7,6 +7,23 @@ using Microsoft.Extensions.Options;
 
 namespace WeihanLi.Web.Middleware
 {
+    public class CustomExceptionHandlerOptions
+    {
+        public Func<HttpContext, ILogger, Exception, Task> OnException { get; set; } =
+            (context, logger, exception) =>
+            {
+                logger.LogError(exception, $"Request exception, requestId: {context.TraceIdentifier}");
+                return Task.CompletedTask;
+            };
+
+        public Func<HttpContext, ILogger, Task> OnRequestAborted { get; set; } =
+            (context, logger) =>
+            {
+                logger.LogInformation($"Request aborted, requestId: {context.TraceIdentifier}");
+                return Task.CompletedTask;
+            };
+    }
+
     public class CustomExceptionHandlerMiddleware
     {
         private readonly RequestDelegate _next;
